Module chess_computer::chess_board

Modules

move 🔒

piece 🔒

Structs

BoardPosition

ChessBoard

ChessBoardPlugin 🔒

GameEndStatusIter

An iterator over the variants of GameEndStatus

PieceColorIter

An iterator over the variants of PieceColor

PieceCreateEvent

Event sent by the ChessBoard to notify that a piece has been placed on the board.

PieceMoveEvent

Event sent by the ChessBoard to notify that a piece has been moved.

PieceTypeIter

An iterator over the variants of PieceType

RequestMoveEvent

Event sent to the ChessBoard to request that a move is made.

ResetBoardEvent

Event sent to the ChessBoard to reset it with the given Fen.

Enums

GameEndStatus

PieceColor

PieceType

Constants

BOARD_SIZE

Functions

game_end_checker 🔒

make_move 🔒

reset_board_state 🔒

setup 🔒